Get live data

+++ db/schema.sql
@@ -30,9 +30,9 @@ CREATE TABLE listings (
   weekly_views INT DEFAULT 0,
   min_stay INT DEFAULT 1,
   max_guests INT,
-  fees INT DEFAULT 0,
+  fees FLOAT DEFAULT 0,
   tax_rate INT,
-  rate INT NOT NULL,
+  rate FLOAT NOT NULL,
   PRIMARY KEY (id),
   FOREIGN KEY (host_id) REFERENCES hosts(id),
   FOREIGN KEY (review_id) REFERENCES reviews(id)
@@ -53,7 +53,7 @@ CREATE TABLE reservations (
   booked_dates_id INT NOT NULL,
   total_adults INT NOT NULL,
   total_pups INT DEFAULT 0,
-  total_charge INT NOT NULL, 
+  total_charge FLOAT NOT NULL, 
   created_at DATETIME DEFAULT CURRENT_TIMESTAMP,
   PRIMARY KEY (id),
   FOREIGN KEY (guest_id) REFERENCES users(id)
